refactor: 분석 모듈 구조 리팩토링 및 taint_analysis.log파일 생성 위치 변경
