+++ gradle.properties
@@ -4,7 +4,7 @@ pluginGroup = io.JoJoonBalSsa.TaintBomb
 pluginName = Taint Bomb auto Java Obfuscator
 pluginRepositoryUrl = https://github.com/JoJoonBalSsa/Taint-Bomb.git
 # SemVer format -> https://semver.org
-pluginVersion = 0.6.4.4
+pluginVersion = 0.6.4.5
 
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 pluginSinceBuild = 233
@@ -18,7 +18,7 @@ platformVersion = 2023.3.8
 # Example: platformPlugins = com.jetbrains.php:203.4449.22, org.intellij.scala:2023.3.27@EAP
 platformPlugins =
 # Example: platformBundledPlugins = com.intellij.java
-platformBundledPlugins =
+platformBundledPlugins = com.intellij.java
 
 # Gradle Releases -> https://github.com/gradle/gradle/releases
 gradleVersion = 8.10.2
