+++ src/main/resources/pyscripts/analysis/utils/resultManager.py
