+++ src/main/resources/pyscripts/create_hash.py
@@ -13,14 +13,21 @@ def calculate_sha256(file_path):
 
 def python_in_directory(directory_path, exclude_file='create_hash.py', output_file='check_hash'):
     """디렉토리 내의 .py 파일들의 SHA-256 해시값을 계산하고 결과를 output_file에 저장합니다."""
+    file_hashes = {}
+
+    # 모든 .py 파일을 탐색하여 해시값 계산
+    for root, _, files in os.walk(directory_path):
+        for file in files:
+            if file.endswith('.py') and file != exclude_file:
+                file_path = os.path.join(root, file)
+                file_hash = calculate_sha256(file_path)
+                file_name = file.replace('.py', '')
+                file_hashes[file_name] = file_hash
+
+    # 딕셔너리를 파일에 저장
     with open(os.path.join(directory_path, output_file), 'w') as output:
-        for root, _, files in os.walk(directory_path):
-            for file in files:
-                if file.endswith('.py') and file != exclude_file:
-                    file_path = os.path.join(root, file)
-                    file_hash = calculate_sha256(file_path)
-                    file_name = file.replace('.py', '')
-                    output.write(f"{file_name} {file_hash}\n")
+        for file_name, file_hash in sorted(file_hashes.items()):
+            output.write(f"{file_name} {file_hash}\n")
     print(f"SHA-256 해시값이 {output_file} 파일에 저장되었습니다.")
